refactor(picker): update to use new step token (#28927)
Issue number: internal

---------

## What is the current behavior?
The picker highlight only uses the `--ion-color-step-150` variable with
a fallback to `#eeeeef`.

## What is the new behavior?
The picker highlight falls back to using the
`--ion-background-color-step-150` variable if it is defined but
prioritizes `--ion-color-step-150` if it is also defined. This is the
new step color token added for high-contrast themes.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

## Other information

No visual diffs are expected.