refactor(segment-button): update to use new step token (#28925)
Issue number: internal

---------

## What is the current behavior?
The segment button checked indicator only uses the
`--ion-color-step-350` variable with a fallback to the
`--ion-background-color`.

## What is the new behavior?
The segment button checked indicator falls back to using the
`--ion-background-color-step-350` variable if it is defined but
prioritizes `--ion-color-step-350` if it is also defined. This is the
new step color token added for high-contrast themes.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

## Other information

No visual diffs are expected.