refactor(angular-server): build with angular 16 (#28901)
[Issue number: Internal

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Ionic 8 drops support for Angular 14 and 15. As a result, Ionic Angular
Server should be built with Angular 16 to account for supported versions
of Angular.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Updates Ionic Angular Server to build with Angular 16

Note: The ng14 and ng15 tests will fail because they are not compatible.
These test apps are being removed as part of FW-5868. However, the rest
of this PR can be reviewed. Once the test apps are removed, I will sync
in the changes and all tests should pass.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/.github/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->


⚠️ Blocked on merging until FW-5868 is complete.